fix: address Copilot review comments on _resolve_plugin, YAML errors, parse_envs
- _resolve_plugin: replace str() token fallback with '' to avoid repr
  of token lists when perform_substitution returns empty
- ParameterFile.evaluate: catch yaml.YAMLError and include file path
  in the re-raised exception for easier debugging
- Move parse_envs to ExecuteProcess (proper base for env semantics);
  Node.parse() delegates to ExecuteProcess.parse_envs, removing
  the duplicated implementation
